third_party/WebKit/LayoutTests/fast/canvas/drawImage-with-broken-image-expected.txt

This test checks behavior of Canvas::drawImage with a broken source image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.drawImage(null, 0, 0) threw exception TypeMismatchError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The 1st argument provided is either null, or an invalid HTMLImageElement object..
PASS ctx.drawImage(null, 0, 0, 20, 20) threw exception TypeMismatchError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The 1st argument provided is either null, or an invalid HTMLImageElement object..
PASS ctx.drawImage(null, 0, 0, 20, 20, 0, 0, 20, 20) threw exception TypeMismatchError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The 1st argument provided is either null, or an invalid HTMLImageElement object..
PASS ctx.drawImage(invalidImage, 0, 0) is undefined
PASS ctx.drawImage(invalidImage, 0, 0, 20, 20) is undefined
PASS ctx.drawImage(invalidImage, 0, 0, 20, 20, 0, 0, 20, 20) is undefined
PASS ctx.drawImage(invalidImage, 0, 0, 0, 20) is undefined
PASS ctx.drawImage(invalidImage, 0, 0, 0, 20, 0, 0, 20, 20) is undefined
PASS successfullyParsed is true

TEST COMPLETE
